Delete işlemindeki sildiği halde status 500  göndermesi düzeltildi.

+++ Devs/src/main/java/Kodlama/io/Devs/Businnes/concretes/ProgrammingLanguageManager.java
@@ -21,35 +21,66 @@ public class ProgrammingLanguageManager implements IProgrammingLanguageService {
 		this.programmingLanguageRepository = programmingLanguageRepository;
 	}
 
+	private boolean isIdExist(int Id) {
+		for(ProgrammingLanguage obje :programmingLanguageRepository.getAll()) {
+			if(obje.getId()==Id) {
+				return true;
+			}
+		}
+		return false;
+	}
+	
+	private boolean isNameExist(String name) {
+		for(ProgrammingLanguage obje :programmingLanguageRepository.getAll()) {
+			if(obje.getName()==name) {
+				return true;
+			}
+		}
+		return false;
+	}
+	
+	private boolean isNameBlank(String name) {
+		return name.isBlank();
+	}
+
 	@Override
 	public List<ProgrammingLanguage> getAll() {
-		
 		return programmingLanguageRepository.getAll();
 	}
 
-	@Override
-	public ProgrammingLanguage getProgrammingLanguageById(int ProgrammingLanguageId) {
-	
-		return programmingLanguageRepository.getProgrammingLanguageById(ProgrammingLanguageId);
-	}
-	
-	@Override
-	public ProgrammingLanguage addProgrammingLanguage(ProgrammingLanguage programmingLanguage) throws Exception {
-	
-		return programmingLanguageRepository.addProgrammingLanguage(programmingLanguage);
-	}
 
 	@Override
-	public void deleteProgrammingLanguageById(int ProgrammingLanguageId) {
-		
+	public ProgrammingLanguage getProgrammingLanguageById(int programmingLanguageId) throws Exception {
+		if(!isIdExist(programmingLanguageId)) {
+			throw new Exception("Girilen :"+ programmingLanguageId +" Id bulunamadı !");
+		}
+		return programmingLanguageRepository.getProgrammingLanguageById(programmingLanguageId);
+	}
+
+
+	@Override
+	public void addProgrammingLanguage(ProgrammingLanguage programmingLanguage) throws Exception {
+		if(isIdExist(programmingLanguage.getId())) {throw new Exception(programmingLanguage.getId()+" ile Ekleme Yapılamaz !"); }
+		if(isNameBlank(programmingLanguage.getName())) {throw new Exception("Boş isim ile Ekleme Yapılamaz !");}
+		if(isNameExist(programmingLanguage.getName())) {throw new Exception(programmingLanguage.getName()+" Tekrar Eklenemez !");}
+		getAll().add(programmingLanguage);
+	}
+
+
+	@Override
+	public void deleteProgrammingLanguageById(int ProgrammingLanguageId) throws Exception {
+		if(!isIdExist(ProgrammingLanguageId)) {throw new Exception("Var olmayan :"+ProgrammingLanguageId +" Id ile silme işlemi Yapamazsınız !");}
 		programmingLanguageRepository.deleteProgrammingLanguageById(ProgrammingLanguageId);
 	}
 
+
 	@Override
-	public ProgrammingLanguage updateProgrammingLanguage(int Id, ProgrammingLanguage programmingLanguage) {
-		
-		return programmingLanguageRepository.updateProgrammingLanguage(Id, programmingLanguage);
+	public void updateProgrammingLanguage(ProgrammingLanguage programmingLanguage)throws Exception {
+		if(!isIdExist(programmingLanguage.getId())) {throw new Exception("Grilen :"+programmingLanguage.getId()+" Id Bulunamadı !");}
+		if(isNameBlank(programmingLanguage.getName())) {throw new Exception("Boş isim ile Ekleme Yapılamaz !");}
+		programmingLanguageRepository. updateProgrammingLanguage(programmingLanguage);		
 	}
 
+	
 
 }
